begin refactoring traversal path
